Fix lossy-conversions lint warnings (#96398)
JDK 20 added a new javac lint warning for possible lossy conversion in compound assignments - because of implicit type casts, e.g.
warning: [lossy-conversions] implicit cast from int to byte in compound assignment is possibly lossy

The change resolves all such warnings, by either widening the type of the left-hand operand, or explicitly casting the type of the right-hand operand.

+++ server/src/test/java/org/elasticsearch/index/snapshots/blobstore/FileInfoTests.java
@@ -168,7 +168,7 @@ public class FileInfoTests extends ESTestCase {
         );
         int numBytes = 0;
         for (int i = 0; i < info.numberOfParts(); i++) {
-            numBytes += info.partBytes(i);
+            numBytes += (int) info.partBytes(i);
         }
         assertEquals(numBytes, 36);
 
@@ -179,7 +179,7 @@ public class FileInfoTests extends ESTestCase {
         );
         numBytes = 0;
         for (int i = 0; i < info.numberOfParts(); i++) {
-            numBytes += info.partBytes(i);
+            numBytes += (int) info.partBytes(i);
         }
         assertEquals(numBytes, 35);
         final int numIters = randomIntBetween(10, 100);
@@ -193,7 +193,7 @@ public class FileInfoTests extends ESTestCase {
             info = new BlobStoreIndexShardSnapshot.FileInfo("foo", metadata, ByteSizeValue.ofBytes(randomIntBetween(1, 1000)));
             numBytes = 0;
             for (int i = 0; i < info.numberOfParts(); i++) {
-                numBytes += info.partBytes(i);
+                numBytes += (int) info.partBytes(i);
             }
             assertEquals(numBytes, metadata.length());
         }
